+++ src/main/java/aero/cubox/door/service/DoorService.java
@@ -1,6 +1,8 @@
 package aero.cubox.door.service;
 
 
+import org.apache.poi.ss.usermodel.Sheet;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -52,5 +54,9 @@ public interface DoorService {
     int getFloorNameVerification(HashMap<String, Object> param);
     int getDoorNameVerification(HashMap<String, Object> param);
 
+    // 엑셀
+    void deleteBuildingAreaFloorDoorAll();
+
+    int addBuildingFloorDoor(Sheet sheet);
 
 }
