+++ src/main/java/aero/cubox/door/service/impl/DoorServiceImpl.java
@@ -3,12 +3,17 @@ package aero.cubox.door.service.impl;
 import aero.cubox.door.service.DoorService;
 import aero.cubox.util.StringUtil;
 import egovframework.rte.fdl.cmmn.EgovAbstractServiceImpl;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import static aero.cubox.door.controller.DoorController.getValue;
 
 @Service("doorService")
 public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorService {
@@ -390,6 +395,176 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
         return doorDAO.getTerminalList(commandMap);
     }
 
+    /**
+     * 엑셀 업로드 시 전체 삭제
+     */
+    @Override
+    @Transactional
+    public void deleteBuildingAreaFloorDoorAll() {
+        doorDAO.deleteDoorAll();
+        doorDAO.deleteFloorAll();
+        doorDAO.deleteAreaAll();
+        doorDAO.deleteBuildingAll();
+    }
 
+    @Override
+    @Transactional
+    public int addBuildingFloorDoor(Sheet sheet) {
+
+        int cnt = 0;
+        HashMap paramMap;
+        HashMap buildingMap = new HashMap();
+        HashMap floorMap = new HashMap();
+        HashMap doorMap = new HashMap();
+
+        // 1. Building
+        for (int i = 1; i <= sheet.getLastRowNum(); i++) {
+            Row row = sheet.getRow(i);
+
+            // 행이 없으면 패스
+            if (row == null) continue;
+
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                               // 빌딩 명
+            String buildingCd = getValue(row.getCell(5)).replaceAll("\n", "<br>");                                               // 빌딩 코드
+
+            if (!buildingMap.containsValue(String.format("%02d", Integer.parseInt(buildingCd)))) { // buildingCd가 buildingMap에 없는 경우
+                buildingMap.put(buildingNm, String.format("%02d", Integer.parseInt(buildingCd)));
+
+                HashMap param = new HashMap();
+                param.put("buildingNm", buildingNm);
+                param.put("buildingCd", String.format("%02d", Integer.parseInt(buildingCd)));
+                param.put("workplaceId", 1);
+
+                addBuilding(param);
+            }
+        }
+
+        // Building List
+        paramMap = new HashMap();
+        List<Map> buildingList = getBuildingList(paramMap);   //빌딩 목록
+
+        // 2. Floor
+        for (int i = 1; i <= sheet.getLastRowNum(); i++) {
+            Row row = sheet.getRow(i);
+
+            if (row == null) continue;
+
+            HashMap<String, String> floorInfo = replaceFloorInfo
+                                (getValue(row.getCell(2)).replaceAll("\n", "<br>")
+                                , getValue(row.getCell(6)).replaceAll("\n", "<br>"));
+
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                            // 빌딩 명
+            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).replaceAll("\n", "<br>")));   // 빌딩 코드
+            String buildingId = getBuildingId(buildingList, buildingNm, buildingCd);                                                                 // 빌딩 id
+            String floorNm = floorInfo.get("floorNm");                                                                                               // 층 명
+            String floorCd = floorInfo.get("floorCd");                                                                                               // 층 코드
+
+            if (!floorMap.containsValue(buildingCd + "_" + floorCd)) {
+                floorMap.put(buildingNm + "_" + floorNm, buildingCd + "_" + floorCd);
+
+                if (!buildingId.equals("")) {
+                    HashMap param = new HashMap();
+                    param.put("floorNm", floorNm);
+                    param.put("floorCd", floorCd);
+                    param.put("buildingId", buildingId);
+                    param.put("buildingCd", buildingCd);
+
+                    addFloor(param);
+                }
+            }
+        }
+
+        // Floor List
+        paramMap = new HashMap();
+        List<HashMap> floorList = getFloorList(paramMap);
+
+        // 3. Door
+        for (int i = 1; i <= sheet.getLastRowNum(); i++) {
+            Row row = sheet.getRow(i);
+
+            if (row == null) continue;
+
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                              // 빌딩 명
+            String doorNm = getValue(row.getCell(3)).replaceAll("\n", "<br>");                                                  // 출입문 명
+            String terminalCd = getValue(row.getCell(4)).replaceAll("\n", "<br>");                                              // 단말기 코드
+            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).replaceAll("\n", "<br>")));     // 빌딩 코드 (2자리로 넣어야함)
+            String floorCd = getValue(row.getCell(6)).replaceAll("\n", "<br>");                                                 // 층 코드 (2자리로 넣어야함)
+            String doorCd = getValue(row.getCell(7)).replaceAll("\n", "<br>");                                                  // 출입문 코드
+            String buildingId = getBuildingId(buildingList, buildingNm, buildingCd);                                                                   // 빌딩 id
+            String floorId = getFloorId(floorList, floorCd, buildingCd);                                                                               // 층 id
+
+            // doorCd 6자리수 변형
+            if (doorCd.length() < 6) {
+                String preNum = "";
+                int num = 6 - doorCd.length();
+                for (int j = 0; j < num; j++) {
+                    preNum += "0";
+                }
+                doorCd = preNum + doorCd;
+            }
+
+            if (!doorMap.containsValue(doorCd)) {
+                doorMap.put(doorNm, doorCd);
+
+                HashMap param = new HashMap();
+                param.put("buildingCd", buildingCd);
+                param.put("floorCd", floorCd);
+                param.put("doorCd", doorCd);
+                param.put("buildingId", buildingId);
+                param.put("areaId", null);
+                param.put("floorId", floorId);
+                param.put("doorNm", doorNm);
+                //  param.put("terminalCd", terminalCd);
+                //  param.put("alarmGroupId", );
+
+                String newDoorId = addDoor(param);
+                if (newDoorId != "") cnt++;
+
+            } else {
+                return -1;
+            }
+        }
+        return cnt;
+    }
+
+    public HashMap replaceFloorInfo(String floorNm, String floorCd) {
+        HashMap floorInfo = new HashMap();
+        floorInfo.put("floorNm", floorNm);
+        floorInfo.put("floorCd", floorCd);
+
+        // floorNm PH층으로 통일
+        if (floorNm.matches(".*PH*.") || floorNm.matches(".*ph*.") || floorNm.matches(".*옥상*.")) {
+            floorInfo.put("floorNm", "PH층");
+        }
+        // floorCd 2자리수 변형
+        if (floorCd.length() == 1) {
+            floorInfo.put("floorCd", "0" + floorCd);
+        }
+        return floorInfo;
+    }
+
+    public String getBuildingId(List<Map> buildingList, String buildingNm, String buildingCd) {
+        String buildingId = "";
+
+        for (int j = 0; j < buildingList.size(); j++) {
+            if (buildingList.get(j).get("building_nm").equals(buildingNm) && buildingList.get(j).get("building_cd").equals(buildingCd)) {
+                buildingId = buildingList.get(j).get("id").toString();
+                break;
+            }
+        }
+        return buildingId;
+    }
+
+    public String getFloorId(List<HashMap> floorList, String floorCd, String buildingCd) {
+        String floorId = "";
+
+        for (int j = 0; j < floorList.size(); j++) {
+            if (floorList.get(j).get("floor_cd").equals(floorCd) && floorList.get(j).get("building_cd").equals(buildingCd)) {
+                floorId = floorList.get(j).get("id").toString();
+                break;
+            }
+        }
+        return floorId;
+    }
 
 }
